+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Unreleased
     - Added Delta TF-IDF Transformer
     - Added document frequency smoothing to TF-IDF Transformer
+    - Added Unform neural net weight initializer
 
 - 0.0.7-alpha
     - Added Support Vector Machine classifier and regressor